Add GitHub Actions workflow to publish documentation
This commit adds a new GitHub Actions workflow file, `.github/workflows/publish.yml`, which will be responsible for automatically publishing the documentation to GitHub Pages. The workflow is triggered on pushes to the `master` branch and can also be manually triggered from the Actions tab. It runs on the latest Ubuntu environment and consists of two steps. The first step checks out the `master` branch, and the second step deploys the documentation using the `mhausenblas/mkdocs-deploy-gh-pages` action. The necessary environment variables, `GITHUB_TOKEN` and `CUSTOM_DOMAIN`, are also set.

Signed-off-by: SammyOina <sammyoina@gmail.com>