Refactor GitHub Actions workflow to use 'main' branch instead of 'master'
The previous GitHub Actions workflow was configured to trigger on the 'master' branch. This commit updates the workflow to trigger on the 'main' branch instead.

This change ensures that the workflow is aligned with the updated branch naming convention. The 'actions/checkout' step has been modified to use the latest version (v4) of the 'actions/checkout' action.

These changes will help maintain consistency and avoid potential confusion in the project's workflow configuration.

Signed-off-by: SammyOina <sammyoina@gmail.com>