NOISSUE - Add GitHub Actions workflow to publish documentation (#18)
* Add GitHub Actions workflow to publish documentation

This commit adds a new GitHub Actions workflow file, `.github/workflows/publish.yml`, which will be responsible for automatically publishing the documentation to GitHub Pages. The workflow is triggered on pushes to the `master` branch and can also be manually triggered from the Actions tab. It runs on the latest Ubuntu environment and consists of two steps. The first step checks out the `master` branch, and the second step deploys the documentation using the `mhausenblas/mkdocs-deploy-gh-pages` action. The necessary environment variables, `GITHUB_TOKEN` and `CUSTOM_DOMAIN`, are also set.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Refactor GitHub Actions workflow to use 'main' branch instead of 'master'

The previous GitHub Actions workflow was configured to trigger on the 'master' branch. This commit updates the workflow to trigger on the 'main' branch instead.

This change ensures that the workflow is aligned with the updated branch naming convention. The 'actions/checkout' step has been modified to use the latest version (v4) of the 'actions/checkout' action.

These changes will help maintain consistency and avoid potential confusion in the project's workflow configuration.

Signed-off-by: SammyOina <sammyoina@gmail.com>

---------

Signed-off-by: SammyOina <sammyoina@gmail.com>