feat: Add CLI command for generating keys
This commit adds a new CLI command for generating public and private key pairs. The keys are used to identify and authenticate users for different agent operations as defined in the computation manifest. The command can be executed by running `./build/cocos-cli keys`. This command generates two files, `private.pem` and `public.pem`, with the private key used with the CLI and the public key added to the computation manifest.

Signed-off-by: SammyOina <sammyoina@gmail.com>